+++ system/activate/set.sh
@@ -1,5 +1,13 @@
 #!/bin/sh
+# shellcheck disable=SC2039
 
-# Set up text editor.
-# Using vi mode instead of emacs by default.
-set -o vi
+# Readline. Currently uses emacs by default.
+# https://unix.stackexchange.com/questions/30454
+case "$EDITOR" in
+    emacs)
+        set -o emacs
+        ;;
+    vi|vim)
+        set -o vi
+        ;;
+esac