+++ system/activate/shell/zsh/bindkey.sh
@@ -1,9 +1,10 @@
 #!/usr/bin/env zsh
 
-# Zsh line editing (zle)
+# Key bindings
+# Updated 2019-08-14.
 
 # Vim mode
 bindkey -v
 
 # Emacs mode
-# bindkey -e
+# > bindkey -e